[CPU Holder Part] Made small stick a bit higher after v2 print
The print was successful, but it's hard to turn the part in place after
inserting it on the CPU Holder.

This is most likely because the small stick is too low, so once the
parts are all inserted the small stick hits the rail or something.

I'm increasing its position by 1mm, while decreasing the height of the
stick by 1mm so the upper edge stays where it is.